adding a schema to the H2 database in order to simulate FIGIS

+++ vme-service/src/test/java/org/fao/fi/figis/dao/FigisDaoTest.java
@@ -40,7 +40,7 @@ public class FigisDaoTest {
 		dao.persist(o);
 		RefVme found = dao.loadRefVme(id);
 		assertNotNull(found);
-		assertEquals(id, found.getId());
+		assertEquals(id, found.getId().intValue());
 
 	}
